Move meta clients to meta package.
