+++ core/client/fs/src/main/java/alluxio/client/meta/RetryHandlingMetaMasterConfigClient.java
@@ -9,7 +9,7 @@
  * See the NOTICE file distributed with this work for information regarding copyright ownership.
  */
 
-package alluxio.client;
+package alluxio.client.meta;
 
 import alluxio.AbstractMasterClient;
 import alluxio.Constants;
