+++ shell/src/main/java/alluxio/cli/fsadmin/command/AbstractFsAdminCommand.java
@@ -12,7 +12,7 @@
 package alluxio.cli.fsadmin.command;
 
 import alluxio.cli.Command;
-import alluxio.client.MetaMasterClient;
+import alluxio.client.meta.MetaMasterClient;
 import alluxio.client.block.BlockMasterClient;
 import alluxio.client.file.FileSystemMasterClient;
