+++ tests/src/test/java/alluxio/server/configuration/ConfigCheckerIntegrationTest.java
@@ -18,7 +18,7 @@ import static org.junit.Assert.assertTrue;
 import alluxio.conf.ServerConfiguration;
 import alluxio.Constants;
 import alluxio.conf.PropertyKey;
-import alluxio.client.MetaMasterClient;
+import alluxio.client.meta.MetaMasterClient;
 import alluxio.grpc.ConfigStatus;
 import alluxio.grpc.Scope;
 import alluxio.multi.process.MultiProcessCluster;
